readonly_backend/requirements.txt
aiofiles==24.1.0
aiomisc==17.3.41
annotated-types==0.6.0
anyio==4.2.0
boto3==1.38.7
botocore==1.38.7
certifi==2025.4.26
click==8.1.7
colorlog==6.8.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.0
fastapi==0.108.0
greenlet==3.0.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.6
jiter==0.9.0
jmespath==1.0.1
openai==1.77.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9.1
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.32.0.post1
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.25.0
